+++ ServiceLayer/Models/GamerModelDto.cs
@@ -2,7 +2,7 @@
 {
     public class GamerModelDto
     {
-        public int GamerId { get; set; }
+        public long GamerId { get; set; }
 
         public string? Gamertag { get; set; }
